Basic double (32bits) everywhere

+++ src/typedefs.h
@@ -4,6 +4,6 @@
 #include <QtGlobal>
 
 typedef uint_fast32_t nFrame;
-typedef double Sample; // Weird : Needs to be 64bit long
+typedef float Sample; // Weird : Needs to be 64bit long
 
 #endif // TYPEDEFS_H
